backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
Pillow==10.1.0